requirements.txt
msal>=1.25.0
requests>=2.31.0
rich>=13.7.0
click>=8.1.0
python-dotenv>=1.0.0
boto3>=1.34.0
azure-mgmt-containerinstance>=10.1.0
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
